Relax Parsing the Commandline
Add a new array
```
static char * command_queue_r[BUFSIZE];
```
where we pass a the commandline to `process_next_command()` in. The commandline here does not contain the linenumber and the checksum anymore, so `process_next_command()` must not care about them.

In `get_command()`:
Only care about the checksumm when we have a linenumber. Otherwise we can't handle a '*' inside the command.

Terminate the commandline with a '\0' at the '*' position.

strtol() has a side effect. It can return the char* just behind the last used digit as the second parameter. Use this to set `command_queue_r[]` when reading the linenumber.

Remove throwing the error for MSG_ERR_NO_LINENUMBER_WITH_CHECKSUM - just ignore it.

Clean a little bug where the checksum was ignored behind a comment.

In `code_has_value()`:
Tolerate leading spaces.

In `process_next_command()`:
Use `command_queue_r[]`.
Drop all that skipping of the linenumber. We did that in `get_command()`.
Drop all that skipping the chars for the codenum to find the start of `current_command_args` - let strtol do the job.
Replace `codenum = code_value_short();` with a direct call to strtol() to have access to the second parameter
`code_is_good` now depends on `current_command_args - 1` that is now the last parsed digit, instead of the first.


Temporary add some debug code.
M113 parses all possible parameters and prints what was read.

# Conflicts:
#	Marlin/Marlin_main.cpp